requirements.txt
speechbrain==0.5.16
pydub
translate
requests==2.27.1
streamlit==1.24.0  
openai 
deepgram-sdk
# git+https://github.com/openai/whisper.git
# git+https://github.com/pyannote/pyannote-audio